requirements-full.txt
pyobjc-core
pyobjc-framework-Cocoa
pyobjc-framework-EventKit
pyobjc-framework-NotificationCenter
ollama
pytesseract
pyautogui
Pillow
sounddevice
faster-whisper
chromadb
sentence-transformers
pymupdf
python-docx
PyYAML
python-dotenv
requests
duckduckgo-search
faiss-cpu
pytest
numpy